==> preprocessing/4-get_city_center_coordinates.ipynb <==
import geopandas as gpd

shape_br = gpd.read_file('../input/shapes/BR_Localidades_2010/BR_Localidades_2010_v1.shp')
shape_br = shape_br[(shape_br['NM_CATEGOR'] == 'CIDADE')][['CD_GEOCODM', 'NM_MUNICIP', 'NM_UF', 'LONG', 'LAT']]
shape_br

shape_br.to_csv('../model/city_center_coordinates.csv', index=False)
