==> preprocessing/5-nightlights_distances_per_city.ipynb <==
import pandas as pd
import numpy as np

nightlights = pd.read_csv('../model/filtered_nighttime_lights.csv')
nightlights

city_center_coordinates = pd.read_csv('../model/city_center_coordinates.csv')
city_center_coordinates

def euclideanDistance(lat, long, code):
    city_center = city_center_coordinates[city_center_coordinates['CD_GEOCODM'] == code].iloc[0]
    if city_center is None:
        return 0
    else:
        return np.sqrt((lat - city_center.LAT)**2 + (long - city_center.LONG)**2)

nightlights['distance'] = nightlights.apply(lambda x: euclideanDistance(x.lat, x.long, x.city_code), axis=1)
nightlights

nightlights.drop(['index_right'], inplace=True, axis=1)


nightlights.rename(columns={'state': 'uf', 'distance': 'distance_to_city' }, inplace=True)


new_nightlights = nightlights[['city_code', 'city', 'uf', 'lat', 'long', 'radiance', 'geometry', 'distance_to_city']]
new_nightlights


new_nightlights.to_csv('../model/coordinates_with_distances.csv', index=False)
