==> preprocessing/6-nearest_nightlights_per_city.ipynb <==
import pandas as pd

df = pd.read_csv('../model/coordinates_with_distances.csv')
df

# remove < 0
df = df[df['radiance'] > 0]
df

df['rank'] = df.groupby('city_code')['distance_to_city'].rank()
nearest = df[df['rank'] <= 100]
nearest

nearest.to_csv('../model/nearest_nightlights_per_city.csv', index=False)
